+++ examples/data-sources/nuodbaas_database/data-source.tf
@@ -1,0 +1,6 @@
+# Get details about a single database
+data "nuodbaas_database" "databaseDetails" {
+  name         = "dbaas"
+  organization = "system"
+  project      = "nuodb"
+}