+++ examples/data-sources/nuodbaas_projects/data-source.tf
@@ -1,0 +1,9 @@
+# Get all projects
+data "nuodbaas_projects" "projectsList" {}
+
+# Get all projects in a given organization
+data "nuodbaas_projects" "projectsList" {
+  filter {
+    organization = "system"
+  }
+}