+++ CHANGELOG.md
@@ -1,3 +1,16 @@
-## 0.1.0 (Unreleased)
+# Changelog
 
-FEATURES:
+## 0.1.0 (Unreleased)
+Initial version of the NuoDB DBaaS provider.
+
+**This is a pre-release version.
+Future versions may not be backwards compatible with this release.**
+
+### Added
+- Provider that can connect to NuoDB Control Plane via its REST API
+- Resource to manage projects
+- Resource to manage databases
+- Data source to list projects
+- Data source to get details about a project
+- Data source to list databases
+- Data source to get details about a database